Move validation function
