+++ src/applications/lgy/coe/form/config/chapters/documents/fileUpload.js
@@ -4,7 +4,7 @@ import { validateFileField } from 'platform/forms-system/src/js/validation';
 
 import FileField from './FileField';
 import UploadRequirements from './UploadRequirements';
-import { validateDocumentDescription } from '../../helpers';
+import { validateDocumentDescription } from '../../../validations';
 import { DOCUMENT_TYPES } from '../../../../status/constants';
 
 const DocumentUploadDescription = () => (
