+++ src/applications/lgy/coe/form/tests/config/helpers.unit.spec.js
@@ -1,10 +1,7 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 import * as helpers from 'platform/forms-system/src/js/helpers';
-import {
-  customCOEsubmit,
-  validateDocumentDescription,
-} from '../../config/helpers';
+import { customCOEsubmit } from '../../config/helpers';
 
 const form = {
   data: {
@@ -52,36 +49,9 @@ afterEach(() => {
   sandbox.restore();
 });
 
-describe('coe helpers', () => {
-  describe('customCOEsubmit', () => {
-    it('should correctly format the form data', () => {
-      sandbox.stub(helpers, 'transformForSubmit').returns(formattedProperties);
-      expect(customCOEsubmit({}, form)).to.equal(result);
-    });
-  });
-  describe('validateDocumentDescription', () => {
-    const errors = addError => [{ attachmentDescription: { addError } }];
-    const fileList = (type, descrip) => [
-      {
-        attachmentType: type,
-        attachmentDescription: descrip,
-      },
-    ];
-    it('should not add an error for non-other type', () => {
-      const spy = sinon.spy();
-      validateDocumentDescription(errors(spy), fileList('ALTA statement', ''));
-      expect(spy.called).to.be.false;
-    });
-    it('should not add an error for Other-type with description', () => {
-      const spy = sinon.spy();
-      validateDocumentDescription(errors(spy), fileList('Other', 'ok'));
-      expect(spy.called).to.be.false;
-    });
-    it('should add an error for a missing attachmentDescription', () => {
-      const spy = sinon.spy();
-      validateDocumentDescription(errors(spy), fileList('Other', ''));
-      expect(spy.called).to.be.true;
-      expect(spy.calledWith('Please provide a description'));
-    });
+describe('customCOEsubmit', () => {
+  it('should correctly format the form data', () => {
+    sandbox.stub(helpers, 'transformForSubmit').returns(formattedProperties);
+    expect(customCOEsubmit({}, form)).to.equal(result);
   });
 });
